Separated brotli and gzip tasks into separate worker files
